+++ packages/babel-macro-util/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.17](https://github.com/fanyonglong/DxWebpack/compare/@dxyl/babel-macro-util@1.0.16...@dxyl/babel-macro-util@1.0.17) (2020-12-03)
+
+**Note:** Version bump only for package @dxyl/babel-macro-util
+
+
+
+
+
 ## [1.0.16](https://github.com/fanyonglong/DxWebpack/compare/@dxyl/babel-macro-util@1.0.15...@dxyl/babel-macro-util@1.0.16) (2020-11-17)
 
 **Note:** Version bump only for package @dxyl/babel-macro-util