+++ packages/webpack/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.16](https://github.com/fanyonglong/DxWebpack/compare/@dxyl/webpack@1.0.15...@dxyl/webpack@1.0.16) (2020-12-03)
+
+**Note:** Version bump only for package @dxyl/webpack
+
+
+
+
+
+
 ## [1.0.15](https://github.com/fanyonglong/DxWebpack/compare/@dxyl/webpack@1.0.14...@dxyl/webpack@1.0.15) (2020-11-29)
 
 **Note:** Version bump only for package @dxyl/webpack